Revert "Initialize CM instance outside of first request (#898)"
This reverts commit c9df0efdb3a54997c21c89eeb11d344fcbaa97a8.
