+++ docker/services/uwsgi.d/50_palace.ini
@@ -6,4 +6,4 @@ pythonpath = %(base)
 
 # python module to import
 module = api.app
-callable = initialize_application()
+callable = app
